New Structure Update
Modulated project structure and added custom account system with AWS
hooks
