+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.4.3.2
+METEOR@1.4.4.2
