+++ .meteor/versions
@@ -1,14 +1,14 @@
-accounts-base@1.2.15
+accounts-base@1.2.17
 accounts-facebook@1.1.1
 accounts-oauth@1.1.15
-accounts-password@1.3.4
+accounts-password@1.3.6
 accounts-ui@1.1.9
 accounts-ui-unstyled@1.2.1
 aldeed:simple-schema@1.5.3
 allow-deny@1.0.5
 autopublish@1.0.7
 autoupdate@1.3.12
-babel-compiler@6.14.1
+babel-compiler@6.18.2
 babel-runtime@1.0.1
 base64@1.0.10
 binary-heap@1.0.10
@@ -22,16 +22,17 @@ callback-hook@1.0.10
 check@1.2.5
 coffeescript@1.0.17
 ddp@1.2.5
-ddp-client@1.3.3
+ddp-client@1.3.4
 ddp-common@1.2.8
 ddp-rate-limiter@1.0.7
-ddp-server@1.3.13
+ddp-server@1.3.14
 deps@1.0.12
 diff-sequence@1.0.7
-ecmascript@0.6.3
+ecmascript@0.7.3
 ecmascript-runtime@0.3.15
+edgee:slingshot@0.7.1
 ejson@1.0.13
-email@1.1.18
+email@1.2.1
 es5-shim@4.6.15
 facebook-config-ui@1.0.0
 facebook-oauth@1.3.0
@@ -53,13 +54,13 @@ mdg:validation-error@0.2.0
 meteor@1.6.1
 meteor-base@1.0.4
 minifier-css@1.2.16
-minifier-js@1.2.18
-minimongo@1.0.21
+minifier-js@2.0.0
+minimongo@1.0.23
 mobile-experience@1.0.4
 mobile-status-bar@1.0.14
-modules@0.7.9
-modules-runtime@0.7.9
-mongo@1.1.16
+modules@0.8.2
+modules-runtime@0.7.10
+mongo@1.1.17
 mongo-id@1.0.6
 npm-bcrypt@0.9.2
 npm-mongo@2.2.24
@@ -69,7 +70,7 @@ observe-sequence@1.0.16
 ordered-dict@1.0.9
 promise@0.8.8
 random@1.0.10
-rate-limit@1.0.7
+rate-limit@1.0.8
 react-meteor-data@0.2.9
 reactive-dict@1.1.8
 reactive-var@1.0.11
@@ -85,17 +86,17 @@ spacebars@1.0.15
 spacebars-compiler@1.1.2
 srp@1.0.10
 standard-minifier-css@1.3.4
-standard-minifier-js@1.2.3
+standard-minifier-js@2.0.0
 templating@1.3.2
 templating-compiler@1.3.2
 templating-runtime@1.3.2
 templating-tools@1.1.2
 tmeasday:check-npm-versions@0.2.0
-tracker@1.1.2
+tracker@1.1.3
 ui@1.0.13
 underscore@1.0.10
 url@1.1.0
 useraccounts:bootstrap@1.14.2
 useraccounts:core@1.14.2
-webapp@1.3.14
+webapp@1.3.15
 webapp-hashing@1.0.9
