+++ client/main.js
@@ -1,20 +1,17 @@
 import React from 'react';
 import { Meteor } from 'meteor/meteor';
 import { render } from 'react-dom';
-import { Router, Route, Link, browserHistory, IndexRedirect} from 'react-router'
+import { Router, Route, browserHistory, IndexRedirect} from 'react-router'
 
 import App from '../imports/App.jsx';
-import Home from '../imports/ui/pages/home.jsx';
-import Profile from '../imports/ui/pages/profile.jsx';
-import Cart from '../imports/ui/pages/cart.jsx';
-import Search from '../imports/ui/pages/search.jsx';
-import Product from '../imports/ui/pages/productView';
-import Submit from '../imports/ui/pages/submit';
-import Login from '../imports/ui/pages/login.jsx'
-import Register from '../imports/ui/components/register.jsx'
-
-import '../imports/startup/client/account-config';
-import '../lib/config/atConfig';
+import Home from '../imports/ui/pages/home/home.jsx';
+import Profile from '../imports/ui/pages/designer/designer.jsx';
+import Cart from '../imports/ui/pages/cart/cart.jsx';
+import Product from '../imports/ui/pages/product/productView.jsx';
+import Submit from '../imports/ui/pages/product/submit';
+import Login from '../imports/ui/pages/login/login.jsx'
+import Register from '../imports/ui/pages/login/register/main.jsx'
+import RecoverPassword from '../imports/ui/pages/login/recoverPassword.jsx'
 
 Meteor.startup(() => {
     render(
@@ -22,14 +19,14 @@ Meteor.startup(() => {
             <Route path="/" component={App}>
                 <IndexRedirect to="/home"/>
                 <Route path="home" component={Home} />
-                <Route path="login" component={Login} />
-                <Route path="register" component={Register} fruits = {"WHAAAAT"}/>
-                <Route path="register/:step" component={Register}/>
-                <Route path="search" component={Search} />
+                <Route path="login" component={Login} noNavBar/>
+                <Route path="recoverPassword" component={RecoverPassword} noNavBar/>
+                <Route path="register" component={Register} noNavBar registration/>
+                <Route path="register/:step" component={Register} noNavBar registration/>
                 <Route path="profile" component={Profile} />
+                <Route path="product/:id" component={Product} />
                 <Route path="cart" component={Cart} />
                 <Route path="submit" component={Submit} />
-                <Route path="product/:entries" component={Product} />
             </Route>
         </Router>
         ,document.getElementById('render-target')
