+++ imports/App.jsx
@@ -2,23 +2,12 @@
  * Created by JohnBae on 12/26/16.
  */
 
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
 import { createContainer } from 'meteor/react-meteor-data';
-
-import {Navbar, NavItem, Nav, FormGroup, FormControl, Button} from 'react-bootstrap'
-import {LinkContainer} from 'react-router-bootstrap'
-import {browserHistory} from 'react-router';
+import NavBar from './ui/components/navBar/navBar.jsx'
 
 import './ui/stylesheets/bootstrap.css'
-import './ui/stylesheets/cart.css'
 import './ui/stylesheets/global.css'
-import './ui/stylesheets/home.css'
-import './ui/stylesheets/navbar.css'
-import './ui/stylesheets/productPreview.css'
-import './ui/stylesheets/productView.css'
-import './ui/stylesheets/search.css'
-import './ui/stylesheets/myStore.css'
-import './ui/stylesheets/login.css'
 
 class App extends Component {
 
@@ -26,71 +15,40 @@ class App extends Component {
         super(props);
     }
 
+    componentDidUpdate(){
+
+        var registering = this.props.children.props.route.registration;
+        console.log("registering?", registering);
+        if(Meteor.user() && !Meteor.user().name && !registering){
+            console.log("LOGGING OUT");
+            Meteor.logout();
+        }
+    }
+
     render() {
-        var self = this;
-        return (
-            <div className = "masterContainer">
-                <NavbarInstance currentUser = {this.props.currentUser} redirect = {this.props.location.pathname}/>
-                {this.props.children}
-            </div>
-        );
+        var showNavBar = !this.props.children.props.route.noNavBar;
+
+        if(showNavBar){
+            return (
+                <div className = "masterContainer">
+                    <NavBar currentUser = {this.props.currentUser}
+                            redirect = {this.props.location.pathname}/>
+                    {this.props.children}
+                </div>
+            );
+        }
+        else{
+            return (
+                <div className = "masterContainer">
+                    {this.props.children}
+                </div>
+            );
+        }
     }
 }
-
-class NavbarInstance extends Component{Z
-
-    logOff(){
-        AccountsTemplates.logout();
-    }
-
-    render(){
-        const {...rest} = this.props;
-        return(
-        <Navbar inverse collapseOnSelect>
-            <Navbar.Header>
-                <Navbar.Brand>
-                    <div className = 'navbar-Logo' onClick={()=>browserHistory.push('/home')}></div>
-                </Navbar.Brand>
-                <Navbar.Toggle />
-            </Navbar.Header>
-            <Navbar.Collapse>
-                <Nav pullRight >
-                    {/*<Navbar.Form pullLeft>
-                        <FormGroup>
-                            <FormControl type="text" placeholder="Search" />
-                        </FormGroup>
-                        {' '}
-                    </Navbar.Form>*/}
-                    <LinkContainer to="/home">
-                        <NavItem>HOME</NavItem>
-                    </LinkContainer>
-                    <LinkContainer to="/profile">
-                        <NavItem>MY STORE</NavItem>
-                    </LinkContainer>
-                    <LinkContainer to="/cart">
-                        <NavItem >CART</NavItem>
-                    </LinkContainer>
-                    {
-                        this.props.currentUser ?
-                            <NavItem onClick = {this.logOff.bind(this)}>PROFILE</NavItem>
-                            :
-                            <LinkContainer to={{pathname: '/login', state: {redirect: this.props.redirect}}}>
-                                <NavItem >LOGIN</NavItem>
-                            </LinkContainer>
-                    }
-                </Nav>
-            </Navbar.Collapse>
-        </Navbar>
-
-        );
-    }
-}
-
-App.propTypes = {
-    currentUser: PropTypes.object,
-};
 
 export default createContainer(() => {
+    Meteor.subscribe("userData");
     return {
         currentUser: Meteor.user(),
     };
