--- imports/startup/client/account-config.js
@@ -1,8 +1,0 @@
-/**
- * Created by JohnBae on 3/28/17.
- */
-import { Accounts } from 'meteor/accounts-base';
-
-Accounts.ui.config({
-    passwordSignupFields: 'USERNAME_AND_EMAIL',
-});
