--- imports/startup/server/accounts/index.js
@@ -1,6 +1,0 @@
-/**
- * Created by JohnBae on 4/6/17.
- */
-
-export * from './email';
-export * from './registration';
