+++ imports/ui/components/navBar/navBar.css
@@ -1,0 +1,6 @@
+#navBar{
+    position: relative;
+    background-color: #000000;
+    width: 100%;
+    height: 100%;
+}
