+++ imports/ui/pages/cart/cart.jsx
@@ -3,10 +3,10 @@
  */
 
 import React, { Component, PropTypes } from 'react';
-import ReactDOM from 'react-dom';
 import {Grid, Row, Col, Panel, Jumbotron, Image, Button, FormGroup, FormControl, ControlLabel, HelpBlock} from 'react-bootstrap'
-import Scroll from 'react-scroll';
-import ProductPreview from '../components/productPreview.jsx';
+import ProductPreview from '../../components/productPreview.jsx';
+
+import './cart.css';
 
 export default class App extends Component {
