+++ imports/ui/pages/home/style.css
@@ -1,6 +1,6 @@
 .homePage {
     position: relative;
-    background: #e4e3e0;
+    background: #FFFFFF;
     padding-left: 50px;
     padding-right: 50px;
 }
