+++ imports/ui/pages/login/register/style.css
@@ -1,0 +1,13 @@
+#basicInfo-profileImg-container{
+    position: relative;
+    width: 100%;
+    height: 300px;
+    background-color: gray;
+}
+
+#basicInfo-profileImg-img{
+    position: relative;
+    height: 100%;
+    max-height: 300px;
+    margin: auto;
+}
