+++ imports/ui/pages/product/submit.js
@@ -6,8 +6,8 @@ import React, { Component, PropTypes } from 'react';
 import ReactDOM from 'react-dom';
 import {Grid, Row, Col, Panel, Jumbotron, Image, Button, FormGroup, FormControl, ControlLabel, Thumbnail} from 'react-bootstrap'
 import Scroll from 'react-scroll';
-import ProductPreview from '../components/productPreview.jsx';
-import DesignerPreview from '../components/designerPreview.jsx'
+import ProductPreview from '../../components/productPreview.jsx';
+import DesignerPreview from '../../components/designerPreview.jsx'
 
 export default class App extends Component {
