--- imports/ui/pages/recoverPassword.js
@@ -1,3 +1,0 @@
-/**
- * Created by JohnBae on 4/12/17.
- */
