+++ imports/ui/stylesheets/global.css
@@ -31,7 +31,7 @@ h1, h2, p, ul, li, a {
     box-shadow: none;
     border: none;
     outline: none;
-    background-color: #303030;
+    background-color: #FFFFFF;
 }
