--- imports/ui/stylesheets/myStore.css
@@ -1,18 +1,0 @@
-.myStorePage{
-    position: relative;
-    background: #e4e3e0;
-    padding-left: 50px;
-    padding-right: 50px;
-}
-
-.myStorePage-about{
-    padding:15px;
-    background: #F4F4F2;
-}
-
-.myStorePage-addItem{
-    height: 150px;
-    width: 150px;
-    border-radius: 150px;
-    border: none;
-}
