--- imports/ui/stylesheets/navbar.css
@@ -1,81 +1,0 @@
-.container {
-    padding: 0px !important;
-}
-
-.navbar-collapse {
-    padding: 0px !important;
-}
-
-.navbar {
-    font-size: 18px;
-    border: none !important;
-    min-height: 35px;
-    padding: 0px !important;
-}
-
-.navbar-Logo {
-    position: relative;
-    float: left !important;
-    margin-top: 10px;
-    top: 10px;
-    display: block;
-    background: url('/assets/revomakeLogo.png') no-repeat center center;
-    background-size: contain;
-    width: 190px;
-}
-
-.navbar-form {
-    padding-right: 15px;
-    margin-top: 0px;
-    margin-bottom: 0px;
-}
-
-.nav {
-    font-size: 18px;
-    padding-top: 20px;
-    margin-bottom: 0;
-    margin-right: 0;
-}
-
-.nav > li > a {
-    position: relative;
-    display: block;
-    padding: 10px 20px;
-}
-
-.navbar-login {
-    width: 50px;
-    height: 50px;
-    border-radius: 50px;
-    border: none;
-}
-
-.navbar-inverse .navbar-nav > li > a {
-    color: #c8c8c4;
-}
-.navbar-inverse .navbar-nav > li > a:hover {
-    color: #f4f4f2;
-}
-.navbar-inverse .navbar-nav > .active > a:focus {
-    color: #313131;
-    background-color: transparent;
-}
-
-.navbar-inverse .navbar-nav > .active {
-    color: #313131;
-    background-color: #e4e3e0;
-}
-
-.navbar-toggle {
-    position: relative;
-    float: right;
-    margin-right: 15px;
-    padding: 9px 10px;
-    margin-top: 15px;
-    margin-bottom: 8px;
-    font-size : 14px;
-    background-color: transparent;
-    background-image: none;
-    border: 0px;
-    border-radius: 4px;
-}
