--- imports/ui/stylesheets/search.css
@@ -1,44 +1,0 @@
-.searchPage{
-    position: relative;
-    background: #e4e3e0;
-    padding-left: 50px;
-    padding-right: 50px;
-}
-
-.searchPage-resultView{
-    width: calc(100% - 300px);
-    padding-top: 15px;
-}
-
-.searchPage-searchTool{
-    width: 300px;
-    height:500px;
-    padding: 10px;
-    background: #F4F4F2;
-    position: fixed;
-    right: 15px;
-    top: 100px;
-}
-
-.searchPage-showBy-text:hover{
-    cursor: pointer;
-    color: #313131;
-}
-
-.searchPage-content {
-    position: relative;
-    width: 100%;
-}
-
-.searchPage-content h1{
-    font-size: 150%;
-    text-transform: capitalize;
-    margin-top: 50px;
-    margin-bottom: 25px;
-    text-decoration: underline;
-}
-
-.searchPage-loadMoreButton{
-    width: 100%;
-    margin: 10px auto;
-}
