--- lib/config/atConfig.js
@@ -1,72 +1,0 @@
-/**
- * Created by JohnBae on 4/4/17.
- */
-
-AccountsTemplates.configure({
-    // Behavior
-    confirmPassword: true,
-    enablePasswordChange: true,
-    forbidClientAccountCreation: false,
-    overrideLoginErrors: true,
-    sendVerificationEmail: false,
-    lowercaseUsername: false,
-    focusFirstInput: true,
-
-    // Appearance
-    showAddRemoveServices: false,
-    showForgotPasswordLink: true,
-    showLabels: true,
-    showPlaceholders: true,
-    showResendVerificationEmailLink: false,
-
-    // Client-side Validation
-    continuousValidation: false,
-    negativeFeedback: false,
-    negativeValidation: true,
-    positiveValidation: true,
-    positiveFeedback: true,
-    showValidating: true,
-
-    // Texts
-    texts: {
-        button: {
-            signUp: "Register Now!"
-        },
-        socialSignUp: "Register",
-        socialIcons: {
-            "meteor-developer": "fa fa-rocket"
-        },
-        title: {
-            forgotPwd: "Recover Your Password"
-        },
-    },
-});
-
-AccountsTemplates.configure({
-    texts: {
-        errors: {
-            loginForbidden: "error.accounts.Please re-enter your account",
-        }
-    }
-});
-
-var pwd = AccountsTemplates.removeField('password');
-AccountsTemplates.removeField('email');
-AccountsTemplates.addFields([
-    {
-        _id: "username",
-        type: "text",
-        displayName: "username",
-        required: true,
-        minLength: 5,
-    },
-    {
-        _id: 'email',
-        type: 'email',
-        required: true,
-        displayName: "email",
-        re: /.+@(.+){2,}\.(.+){2,}/,
-        errStr: 'Invalid email',
-    },
-    pwd
-]);
