+++ package.json
@@ -8,13 +8,14 @@
     "babel-runtime": "^6.20.0",
     "bcrypt": "^1.0.1",
     "classnames": "^2.2.5",
+    "draft-js": "^0.10.0",
     "fs-jetpack": "^0.10.5",
     "immutable": "^3.8.1",
     "meteor-node-stubs": "~0.2.0",
-    "react": "^15.4.2",
+    "react": "^15.5.4",
     "react-addons-pure-render-mixin": "^15.4.1",
     "react-bootstrap": "^0.30.8",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.5.4",
     "react-redux": "^5.0.3",
     "react-router": "^3.0.0",
     "react-router-bootstrap": "^0.23.1",
